+++ packages/tarijs/vitest.config.ts
@@ -1,8 +1,8 @@
-import { defineConfig } from 'vitest/config';
-import { loadEnv } from 'vite';
+import { defineConfig } from "vitest/config";
+import { loadEnv } from "vite";
 
 export default defineConfig({
-  root: '.',
+  root: ".",
   test: {
     env: loadEnv("", "./integration-tests/", ""),
   },
